fix : fix sms exception

+++ Zelut/Api/Zelut.Application/DTOs/Sms/SmsAsanakResult.cs
@@ -3,7 +3,7 @@
 public class SmsAsanakResult
 {
     public Meta meta { get; set; }
-    public List<int> data { get; set; }
+    public List<long> data { get; set; }
 }
 
 public class Meta
